+++ lib/widgets/homeWidgets/catalog_image.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_catalog/widgets/themes.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogImage extends StatelessWidget {
+  final String catalogImage;
+  const CatalogImage({Key? key, required this.catalogImage}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Image.network(catalogImage)
+        .box
+        .color(MyTheme.creamColor)
+        .make()
+        .p16()
+        .w40(context);
+  }
+}
