+++ lib/widgets/themes.dart
@@ -18,6 +18,7 @@ class MyTheme {
         brightness: Brightness.dark,
       );
 
+  // static Color creamColor = Color(0xfff5f5f5);
   static Color creamColor = Color(0xfffae8e0);
   static Color darkBlueColor = Color(0xff000075);
 }
